chore: make utils ts and refactor tests
